# make ResourceLimiter operate on Store data; add hooks for entering and exiting native code