--- a/tests/all/main.rs
+++ b/tests/all/main.rs
@@ -20,6 +20,7 @@ mod module;
 mod module_linking;
 mod module_serialize;
 mod name;
+mod native_hooks;
 mod pooling_allocator;
 mod stack_overflow;
 mod store;