Based on #2897 - draft until that PR lands.

This PR does two things:

* The `ResourceLimiter` trait is now implemented on the data inside a `Store` rather than as a standalone owned type. This change required changing the structure of the Store internals a bunch. Additionally, the trait is now defined in `wasmtime_runtime` and re-exported by `wasmtime`.
* Two new functions, `entering_native_code_hook` and `exiting_native_code_hook`, are added to `Store`. These functions get to operate on the data inside a `Store` much like resource limiter's mut methods. These functions give embedding the opportunity to measure time (or other resources) and trap before entering any native code (i.e. a `Func`) linked into the `Store`, and again before returning from that native code.

<!--

Please ensure that the following steps are all taken care of before submitting
the PR.

- [ ] This has been discussed in issue #..., or if not, please tell us why
  here.
- [ ] A short description of what this does, why it is needed; if the
  description becomes long, the matter should probably be discussed in an issue
  first.
- [ ] This PR contains test cases, if meaningful.
- [ ] A reviewer from the core maintainer team has been assigned for this PR.
  If you don't know who could review this, please indicate so. The list of
  suggested reviewers on the right can help you.

Please ensure all communication adheres to the [code of
conduct](https://github.com/bytecodealliance/wasmtime/blob/master/CODE_OF_CONDUCT.md).
-->